requirements.txt

scikit-learn
scikit-surprise
streamlit
pandas
numpy
